refactor: starting to clean up `util` tests
