+++ test/functionality/dataflow/elements/functions/function-call.ts
@@ -1,4 +1,4 @@
-import { assertDataflow, withShell } from '../../../helper/shell'
+import { assertDataflow, withShell } from '../../../_helper/shell'
 import { DataflowGraph, EdgeType, initializeCleanEnvironments } from '../../../../../src/dataflow'
 import { define, popLocalEnvironment, pushLocalEnvironment } from '../../../../../src/dataflow/environments'
 import { UnnamedArgumentPrefix } from '../../../../../src/dataflow/internal/process/functions/argument'
