+++ test/functionality/r-bridge/lang/ast/parse-expression-lists.ts
@@ -1,5 +1,5 @@
-import { assertAst, withShell } from '../../../helper/shell'
-import { exprList, numVal } from '../../../helper/ast-builder'
+import { assertAst, withShell } from '../../../_helper/shell'
+import { exprList, numVal } from '../../../_helper/ast-builder'
 import { rangeFrom } from '../../../../../src/util/range'
 import { RType } from '../../../../../src/r-bridge'
