+++ test/functionality/r-bridge/lang/ast/parse-pipes.ts
@@ -1,5 +1,5 @@
-import { assertAst, withShell } from '../../../helper/shell'
-import { exprList } from '../../../helper/ast-builder'
+import { assertAst, withShell } from '../../../_helper/shell'
+import { exprList } from '../../../_helper/ast-builder'
 import { rangeFrom } from '../../../../../src/util/range'
 import { RType } from '../../../../../src/r-bridge'
 import { MIN_VERSION_PIPE } from '../../../../../src/r-bridge/lang-4.x/ast/model/versions'
