+++ test/functionality/r-bridge/lang/ast/parse-snippets.ts
@@ -1,5 +1,5 @@
-import { exprList, numVal } from '../../../helper/ast-builder'
-import { assertAst, withShell } from '../../../helper/shell'
+import { exprList, numVal } from '../../../_helper/ast-builder'
+import { assertAst, withShell } from '../../../_helper/shell'
 import { rangeFrom } from '../../../../../src/util/range'
 import { RType } from '../../../../../src/r-bridge'
 import { ensureExpressionList } from '../../../../../src/r-bridge/lang-4.x/ast/parser/xml/internal'
