+++ test/functionality/r-bridge/sessions.ts
@@ -1,8 +1,8 @@
 import chai, { assert } from 'chai'
 import fs from 'fs'
 import { randomString } from '../../../src/util/random'
-import { testRequiresNetworkConnection } from '../helper/network'
-import { testWithShell, withShell } from '../helper/shell'
+import { testRequiresNetworkConnection } from '../_helper/network'
+import { testWithShell, withShell } from '../_helper/shell'
 import { isInstallTest } from '../main.spec'
 import { parseCSV } from '../../../src/r-bridge'
 import { log, LogLevel } from '../../../src/util/log'
