+++ test/functionality/statistics/features/control-flow.ts
@@ -1,4 +1,4 @@
-import { withShell } from '../../helper/shell'
+import { withShell } from '../../_helper/shell'
 import { testForFeatureForInput } from '../statistics.spec'
 import { RFalse, RTrue } from '../../../../src/r-bridge'
