+++ test/functionality/statistics/features/defined-functions.ts
@@ -1,4 +1,4 @@
-import { withShell } from '../../helper/shell'
+import { withShell } from '../../_helper/shell'
 import { testForFeatureForInput } from '../statistics.spec'
 import { MIN_VERSION_LAMBDA } from '../../../../src/r-bridge/lang-4.x/ast/model/versions'
