+++ test/functionality/util/objects.ts
