+++ test/functionality/util/strings.ts
